+++ prom2teams/message/parser.py
@@ -5,11 +5,14 @@ import logging
 logger = logging.getLogger()
 
 def check_fields(json_alerts_attr, json_alerts_labels_attr, json_alerts_annotations_attr):
-    mandatory_fields = ['alertname', 'status', 'instance', 'summary']
-    optional_fields = ['severity', 'description']
+    mandatory_fields = ['alertname', 'status', 'summary']
+    optional_fields = ['severity', 'description', 'instance']
     fields = mandatory_fields + optional_fields
 
     alert_fields = {}
+    
+    # Set the instance to 'none' by default. 
+    alert_fields['alert_instance'] = 'none'
 
     for field in fields:
         alert_field_key = 'alert_' + field